Move Emitter helper files to `Support` folder
